Added multipart loss compute.

+++ hungarian_loss/const.py
@@ -2,7 +2,7 @@
 
 import tensorflow as tf
 
-ZERO = tf.constant(0, tf.float16)
+ZERO = tf.constant(0, tf.float32)
 """The constant for 0."""
-ONE = tf.constant(1, tf.float16)
+ONE = tf.constant(1, tf.float32)
 """The constant for 1."""
